refactor: split session store into different composables
