+++ frontend/app/src/components/account-management/login/LoginForm.vue
@@ -6,7 +6,6 @@ import { toMessages } from '@/utils/validation';
 import { deleteBackendUrl, getBackendUrl, saveBackendUrl } from '@/utils/account-management';
 import { compareTextByKeyword } from '@/utils/assets';
 import { useSessionAuthStore } from '@/store/session/auth';
-import { useSessionStore } from '@/store/session';
 import { useDynamicMessages } from '@/composables/dynamic-messages';
 import { useInterop } from '@/composables/electron-interop';
 import { useUsersApi } from '@/composables/api/session/users';
@@ -14,6 +13,7 @@ import WelcomeMessageDisplay from '@/components/account-management/login/Welcome
 import IncompleteUpgradeAlert from '@/components/account-management/login/IncompleteUpgradeAlert.vue';
 import PremiumSyncConflictAlert from '@/components/account-management/login/PremiumSyncConflictAlert.vue';
 import ExternalLink from '@/components/helper/ExternalLink.vue';
+import { useLogout } from '@/modules/account/use-logout';
 import type { LoginCredentials, SyncApproval } from '@/types/login';
 
 const props = withDefaults(
@@ -48,7 +48,7 @@ const touched = () => emit('touched');
 const newAccount = () => emit('new-account');
 const backendChanged = (url: string | null) => emit('backend-changed', url);
 
-const { logoutRemoteSession } = useSessionStore();
+const { logoutRemoteSession } = useLogout();
 
 const username = ref<string>('');
 const usernameSearch = ref<string>('');
