+++ frontend/app/src/components/dashboard/SnapshotActionButton.vue
@@ -4,13 +4,13 @@ import SnapshotImportDialog from '@/components/dashboard/SnapshotImportDialog.vu
 import { useMessageStore } from '@/store/message';
 import { useStatisticsStore } from '@/store/statistics';
 import { usePeriodicStore } from '@/store/session/periodic';
-import { useSessionStore } from '@/store/session';
 import { useInterop } from '@/composables/electron-interop';
 import { useSnapshotApi } from '@/composables/api/settings/snapshot-api';
 import { useBalances } from '@/composables/balances';
 import { usePremium } from '@/composables/premium';
 import DateDisplay from '@/components/display/DateDisplay.vue';
 import MenuTooltipButton from '@/components/helper/MenuTooltipButton.vue';
+import { useLogout } from '@/modules/account/use-logout';
 import type { Writeable } from '@rotki/common';
 import type { AllBalancePayload } from '@/types/blockchain/accounts';
 
@@ -23,7 +23,7 @@ const importSnapshotDialog = ref<boolean>(false);
 
 const { t } = useI18n();
 const premium = usePremium();
-const { logout } = useSessionStore();
+const { logout } = useLogout();
 const { lastBalanceSave } = storeToRefs(usePeriodicStore());
 const { fetchBalances } = useBalances();
 const { getPath } = useInterop();
